UHF-8650: Use MediaEntityBundle base class to minimize the duplicated code.
